module13/requirements.txt
aiosmtplib==2.0.2
annotated-types==0.6.0
anyio==3.7.1
blinker==1.7.0
click==8.1.7
dnspython==2.4.2
email-validator==2.1.0.post1
fastapi==0.104.1
fastapi-mail==1.4.1
h11==0.14.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
pydantic==2.5.0
pydantic-settings==2.0.3
pydantic_core==2.14.1
python-dotenv==1.0.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.24.0.post1
